Synchronize with main branch

+++ zh-Hant/codes/c/chapter_computational_complexity/time_complexity.c
@@ -90,7 +90,7 @@ int expRecur(int n) {
 }
 
 /* 對數階（迴圈實現） */
-int logarithmic(float n) {
+int logarithmic(int n) {
     int count = 0;
     while (n > 1) {
         n = n / 2;
@@ -100,14 +100,14 @@ int logarithmic(float n) {
 }
 
 /* 對數階（遞迴實現） */
-int logRecur(float n) {
+int logRecur(int n) {
     if (n <= 1)
         return 0;
     return logRecur(n / 2) + 1;
 }
 
 /* 線性對數階 */
-int linearLogRecur(float n) {
+int linearLogRecur(int n) {
     if (n <= 1)
         return 1;
     int count = linearLogRecur(n / 2) + linearLogRecur(n / 2);
