+++ zh-Hant/codes/csharp/chapter_computational_complexity/time_complexity.cs
@@ -10,7 +10,7 @@ public class time_complexity {
     void Algorithm(int n) {
         int a = 1;  // +0（技巧 1）
         a += n;  // +0（技巧 1）
-                    // +n（技巧 2）
+        // +n（技巧 2）
         for (int i = 0; i < 5 * n + 1; i++) {
             Console.WriteLine(0);
         }
@@ -118,7 +118,7 @@ public class time_complexity {
     }
 
     /* 對數階（迴圈實現） */
-    int Logarithmic(float n) {
+    int Logarithmic(int n) {
         int count = 0;
         while (n > 1) {
             n /= 2;
@@ -128,13 +128,13 @@ public class time_complexity {
     }
 
     /* 對數階（遞迴實現） */
-    int LogRecur(float n) {
+    int LogRecur(int n) {
         if (n <= 1) return 0;
         return LogRecur(n / 2) + 1;
     }
 
     /* 線性對數階 */
-    int LinearLogRecur(float n) {
+    int LinearLogRecur(int n) {
         if (n <= 1) return 1;
         int count = LinearLogRecur(n / 2) + LinearLogRecur(n / 2);
         for (int i = 0; i < n; i++) {
@@ -181,12 +181,12 @@ public class time_complexity {
         count = ExpRecur(n);
         Console.WriteLine("指數階（遞迴實現）的運算量 = " + count);
 
-        count = Logarithmic((float)n);
+        count = Logarithmic(n);
         Console.WriteLine("對數階（迴圈實現）的運算量 = " + count);
-        count = LogRecur((float)n);
+        count = LogRecur(n);
         Console.WriteLine("對數階（遞迴實現）的運算量 = " + count);
 
-        count = LinearLogRecur((float)n);
+        count = LinearLogRecur(n);
         Console.WriteLine("線性對數階（遞迴實現）的運算量 = " + count);
 
         count = FactorialRecur(n);
