+++ zh-Hant/codes/python/chapter_computational_complexity/time_complexity.py
@@ -77,7 +77,7 @@ def exp_recur(n: int) -> int:
     return exp_recur(n - 1) + exp_recur(n - 1) + 1
 
 
-def logarithmic(n: float) -> int:
+def logarithmic(n: int) -> int:
     """對數階（迴圈實現）"""
     count = 0
     while n > 1:
@@ -86,14 +86,14 @@ def logarithmic(n: float) -> int:
     return count
 
 
-def log_recur(n: float) -> int:
+def log_recur(n: int) -> int:
     """對數階（遞迴實現）"""
     if n <= 1:
         return 0
     return log_recur(n / 2) + 1
 
 
-def linear_log_recur(n: float) -> int:
+def linear_log_recur(n: int) -> int:
     """線性對數階"""
     if n <= 1:
         return 1
